Refactor: Correct permission class name 'to IsReviewerOrAssigneeOrAdmin and update UserRegestration model to remove unnecessary fields

+++ task_app/api/views.py
@@ -1,8 +1,8 @@
+from asyncio import tasks
+
 from rest_framework.views import APIView
 from rest_framework.permissions import AllowAny, IsAuthenticated
-
-from auth_app.api import serializers
-from .permissions import IsRieviewerOrAssigneeOrAdmin
+from .permissions import IsReviewerOrAssigneeOrAdmin
 from .serializer import TaskSerializer, CommentsSerializer
 from rest_framework.authtoken.models import Token
 from rest_framework.response import Response
@@ -31,7 +31,7 @@ class TaskListView(APIView):
 
 class TaskDetailView(APIView):
         
-        permission_classes = [IsRieviewerOrAssigneeOrAdmin, IsAuthenticated]
+        permission_classes = [IsReviewerOrAssigneeOrAdmin, IsAuthenticated]
 
         def get(self, request, pk):
             tasks = CreateTask.objects.get(pk=pk)
@@ -55,7 +55,7 @@ class TaskDetailView(APIView):
 
 class CommentView(APIView):
      
-    permission_classes = [IsRieviewerOrAssigneeOrAdmin, IsAuthenticated]
+    permission_classes = [IsReviewerOrAssigneeOrAdmin, IsAuthenticated]
 
     def get(self, request, pk):
             comment = Comments.objects.filter(task=pk)
@@ -72,7 +72,7 @@ class CommentView(APIView):
      
 class CommentDetailView(APIView):
      
-    permission_classes = [IsRieviewerOrAssigneeOrAdmin, IsAuthenticated]
+    permission_classes = [IsReviewerOrAssigneeOrAdmin, IsAuthenticated]
 
     def get(self, request, task_pk, comment_pk):
         comment = Comments.objects.get(task=task_pk, pk=comment_pk)
@@ -80,6 +80,7 @@ class CommentDetailView(APIView):
         return Response(serializer.data)
     
     def delete(self, request, task_pk, comment_pk):
+        self.check_object_permissions(request, tasks)
         comment = Comments.objects.get(task=task_pk, pk=comment_pk)  
         comment.delete()
         return Response({"message" : "successfully deleted",}, status=204)
@@ -87,7 +88,7 @@ class CommentDetailView(APIView):
 
 class AssignToMeView(APIView):
 
-    permission_classes = [IsRieviewerOrAssigneeOrAdmin, IsAuthenticated]
+    permission_classes = [IsAuthenticated]
     def get(self, request):
         tasks = CreateTask.objects.filter(assignee_id=request.user)
         serializer = TaskSerializer(tasks, many=True)
@@ -96,7 +97,7 @@ class AssignToMeView(APIView):
 
 class ReviewingView(APIView):
 
-    permission_classes = [IsRieviewerOrAssigneeOrAdmin, IsAuthenticated]
+    permission_classes = [IsAuthenticated]
     def get(self, request):
         tasks = CreateTask.objects.filter(reviewer_id=request.user)
         serializer = TaskSerializer(tasks, many=True)
